study buddy/backend/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
murf-python-sdk
python-dotenv